requirements.txt
Django==1.5.1
PIL==1.1.7
South==0.8.1
argparse==1.2.1
django-braces==1.2.1
django-debug-toolbar==0.9.4
django-extensions==1.3.3
django-model-utils==1.4.0
django-social-auth==0.7.25
django-widget-tweaks==1.3
httplib2==0.8
ipython==1.1.0
oauth2==1.5.211
python-openid==2.2.5
python-twitter==1.0
selenium==2.33.0
simplejson==3.3.0
six==1.5.2
wsgiref==0.1.2